修复自动部署工具

+++ SixpenceStudio.Core/Utils/FileUtil.cs
@@ -94,7 +94,10 @@ namespace SixpenceStudio.Core.Utils
                     }
                     else
                     {
-                        DeleteFile(i.FullName);
+                        if (!ignoreList.Contains(i.Name))
+                        {
+                            DeleteFile(i.FullName);
+                        }
                     }
                 }
             }
